Import my original ad hoc scripts.
`create_overlays.sh` has until now been maintained by directly editing `/usr/local/bin/create_overlays.sh` on my server machine. This made sense in the early days when it was much simpler, but it probably should have moved to git a long time ago.

`cow-status` is a one-liner to show which overlays are almost full. It was historically maintained in a similar way. This functionality should be rolled into the main script, probably.

These script are currently very specific to my setup and not designed to be user-friendly. Subsequent commits with replace them with something nicer, but this commit exists to show the history.